Summary
The front end files here have minor changes to allow for redirecting and php additions for the University dropdowns mostly.

The backend files are as follows:
create event: still in progress, not updating the database.
getMyEvents/getMyRsos: was working, but gets an error.
getUniversity will get the universityID by name but I can't test this due to a problem adding universities to the database.